Added Hazel camera, all working. about to optimise rayDirection recall

+++ RT2App/src/Colour.h
@@ -10,15 +10,15 @@
 
 void write_colour(char* out, colour pixel_colour, int& buffer_position, int samples_per_pixel) {
 
-	double scale = 1.0 / samples_per_pixel;
+	float scale = 1.0f / samples_per_pixel;
 
-	double r = sqrt(pixel_colour.x * scale);
-	double g = sqrt(pixel_colour.y * scale);
-	double b = sqrt(pixel_colour.z * scale);
+	float r = sqrt(pixel_colour.x * scale);
+	float g = sqrt(pixel_colour.y * scale);
+	float b = sqrt(pixel_colour.z * scale);
 
-	uint32_t ir = static_cast<uint32_t>(256 * clamp(r, 0.0, 0.999));
-	uint32_t ig = static_cast<uint32_t>(256 * clamp(g, 0.0, 0.999));
-	uint32_t ib = static_cast<uint32_t>(256 * clamp(b, 0.0, 0.999));
+	uint32_t ir = static_cast<uint32_t>(256 * clamp(r, 0.0f, 0.999f));
+	uint32_t ig = static_cast<uint32_t>(256 * clamp(g, 0.0f, 0.999f));
+	uint32_t ib = static_cast<uint32_t>(256 * clamp(b, 0.0f, 0.999f));
 
 	//out << ir << ' ' << ig << ' ' << ib << '\n';
